+++ docs/viem-account.md
@@ -1,0 +1,17 @@
+
+```
+async function haloSignDigest(address: Hex, digest: HexString): Promise<HaloResSign> {
+    return await execHaloCmdWeb({
+        digest,
+        keyNo: 8,
+        password: '000000',
+    })
+}
+
+function createHaloAccount(address) {
+    return createExternalSignerAccount(
+        address,
+        async (digest: Hex): Promise<Hex> => await haloSignDigest(address, digest)
+    )
+}
+```
